refactoring to make it easier to read
