+++ habit_tracker/Habit/Habit.cs
@@ -1,0 +1,11 @@
+ 
+ 
+ namespace Habit{
+ public class DrinkingWater
+    {
+        public int Id { get; set; }
+        public int Quantity { get; set; }
+        public string HabitType { get; set; }
+        public DateTime Date { get; set; }
+    }
+ }
