+++ habit_tracker/Program.cs
@@ -1,390 +1,10 @@
 ﻿
-using System;
-using System.Dynamic;
-using System.Globalization;
+using Services;
 
-using Microsoft.Data.Sqlite;
-
-namespace habit_tracker
+class Program
 {
-
-    public class DrinkingWater
+    static void Main(string[] args)
     {
-        public int Id { get; set; }
-        public int Quantity { get; set; }
-        public string HabitType { get; set; }
-        public DateTime Date { get; set; }
-    }
-    class Program
-    {
-        private static string dbFile;
-        private static string connectionString;
-
-        static void Main(string[] args)
-        {
-            // main function creates database on start up if not exists, in same folder
-            dbFile = "./habit_tracker.db";
-            connectionString = $"Data Source={dbFile}";
-            using (var connection = new SqliteConnection(connectionString))
-            {
-
-                connection.Open();
-                var tableCmd = connection.CreateCommand();
-
-                tableCmd.CommandText = @"CREATE TABLE IF NOT EXISTS drinking_water (
-                Id INTEGER PRIMARY KEY AUTOINCREMENT,
-                Date TEXT,
-                HabitType TEXT,
-                Quantity INTEGER
-                )";
-
-                tableCmd.ExecuteNonQuery();
-
-
-                connection.Close();
-
-
-            }
-            //entry point for menu
-            GetUserInput();
-
-        }
-
-        static void GetUserInput()
-        {
-            // using switch statement for menu
-            bool closeApp = false;
-            while (!closeApp)
-            {
-                Console.Clear();
-                Console.WriteLine("\nMAIN MENU");
-                Console.WriteLine("\nPlease Enter your Choice!");
-                Console.WriteLine("\nType 0 to close Application");
-                Console.WriteLine("Enter 1 to View All Records.");
-                Console.WriteLine("Enter 2 to Insert Record.");
-                Console.WriteLine("Enter 3 to Delete Record.");
-                Console.WriteLine("Enter 4 to Update Record.");
-                Console.WriteLine("-----------------------------------------\n");
-
-                string? command = Console.ReadLine();
-
-                switch (command)
-                {
-                    case "0":
-                        closeApp = true;
-                        break;
-                    case "1":
-                        GetAllRecords();
-                        break;
-                    case "2":
-                        Insert();
-                        break;
-                    case "3":
-                        Delete();
-                        break;
-                    case "4":
-                        Update();
-                        break;
-                    default:
-                        Console.WriteLine("not a valid command.Please choose from 0 to 4");
-                        break;
-
-                }
-            }
-        }
-        private static void Insert()
-        {
-            string? date = GetDateInput();
-            string? habit = GetHabit();
-            int Quantity = GetNumberInput("\n\nPlease Insert a number of glasses or other measure of your choice (no decimals allowed)\n\n");
-            // wrapped in try statement incase error accessing database
-            try
-            {
-                using (var connection = new SqliteConnection(connectionString))
-                {
-                    connection.Open();
-                    var tableCmd = connection.CreateCommand();
-                    tableCmd.CommandText = "INSERT INTO drinking_water(Date, HabitType, Quantity) VALUES(@date, @habit, @Quantity)";
-                    tableCmd.Parameters.AddWithValue("@date", date);
-                    tableCmd.Parameters.AddWithValue("@habit", habit);
-                    tableCmd.Parameters.AddWithValue("@Quantity", Quantity);
-
-                    tableCmd.ExecuteNonQuery();
-                    connection.Close();
-                }
-            }
-            catch (SqliteException ex)
-            {
-                Console.WriteLine("Unexpected Database Error");
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine("Unexpected Error");
-            }
-
-        }
-
-        internal static int GetNumberInput(string message)
-        {
-            int amount = 0;
-            Console.WriteLine(message);
-
-            string? numberInput = Console.ReadLine();
-            numberInput = numberInput.Trim();
-            while (numberInput == null || !int.TryParse(numberInput, out amount) || amount < 0)
-            {
-                Console.WriteLine("that was an incorrect number");
-                numberInput = Console.ReadLine();
-            }
-            return amount;
-        }
-        internal static string? GetDateInput()
-        {
-            Console.WriteLine("\nEnter the date (dd-MM-yy).");
-            Console.WriteLine("Type 0 to return, or T to use today's date.");
-
-            string? dateInput = Console.ReadLine();
-            // adds todays date if T is pressed
-            if (dateInput.Equals("T", StringComparison.OrdinalIgnoreCase))
-                return DateTime.Now.ToString("dd-MM-yy");
-
-            if (dateInput == "0") return null;
-
-            while (!DateTime.TryParseExact(
-                dateInput,
-                "dd-MM-yy",
-                CultureInfo.InvariantCulture,
-                DateTimeStyles.None,
-                out _
-            ))
-            {
-                Console.WriteLine("Wrong format. Use dd-MM-yy.\nType 0 to return.");
-
-                dateInput = Console.ReadLine();
-
-                if (dateInput == "0")
-                    return null;
-            }
-
-            return dateInput;
-        }
-
-        internal static string GetHabit()
-        {
-            Console.WriteLine("Please enter the type, Max 20 characters long");
-            string? typeInput = Console.ReadLine();
-            while (typeInput == null || typeInput.Length > 20)
-            {
-                Console.WriteLine("Please enter the type, Max 20 characters long");
-                typeInput = Console.ReadLine();
-            }
-
-            return typeInput;
-        }
-
-        // Printing function
-        private static void PrintAllRecords()
-        {
-
-            Console.Clear();
-            try
-            {
-                using (var connection = new SqliteConnection(connectionString))
-                {
-                    connection.Open();
-                    var tableCmd = connection.CreateCommand();
-                    tableCmd.CommandText = $"SELECT * FROM drinking_water";
-
-                    List<DrinkingWater> tableData = new();
-                    SqliteDataReader reader = tableCmd.ExecuteReader();
-                    if (reader.HasRows)
-                    {
-                        while (reader.Read())
-                        {
-                            tableData.Add(
-                                new DrinkingWater
-                                {
-                                    Id = reader.GetInt32(0),
-                                    Date = DateTime.ParseExact(reader.GetString(1), "dd-MM-yy", CultureInfo.InvariantCulture),
-                                    HabitType = reader.GetString(2),
-                                    Quantity = reader.GetInt32(3)
-                                }
-                            );
-                        }
-                    }
-                    else
-                    {
-                        Console.WriteLine("No Rows Found, DataBase is Empty");
-                    }
-                    connection.Close();
-                    Console.WriteLine("ID--Type--Date--Amount----------");
-                    Console.WriteLine("---------------------------------");
-                    foreach (var dw in tableData)
-                    {
-                        Console.WriteLine($"{dw.Id} - {dw.HabitType} - {dw.Date.ToString("dd-MM-yy")} - Quantity: {dw.Quantity} ");
-                    }
-                    Console.WriteLine("---------------------------------");
-                }
-            }
-            catch (SqliteException ex)
-            {
-                Console.WriteLine("Unexpected Database Error");
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine("Unexpected Error");
-            }
-        }
-
-        private static void GetAllRecords()
-        {
-            PrintAllRecords();
-            Console.WriteLine("\nPress any key to return to the menu...");
-            Console.ReadKey(true);
-        }
-
-        private static void Delete()
-        {
-            bool keepDeleting = true;
-
-            while (keepDeleting)
-            {
-                PrintAllRecords();
-
-                Console.WriteLine("Enter the ID of the habit you wish to delete, or type 'x' to return to the menu:");
-                string? input = Console.ReadLine();
-
-                if (input == null) continue;
-
-                if (input.ToLower() == "x")
-                {
-                    keepDeleting = false;
-                    break;
-                }
-                int recordId;
-                if (!int.TryParse(input, out recordId))
-                {
-                    Console.WriteLine("Invalid input. Please enter a number or 'x' to exit.");
-                    continue;
-                }
-                try
-                {
-                    using (var connection = new SqliteConnection(connectionString))
-                    {
-                        connection.Open();
-                        var tableCmd = connection.CreateCommand();
-                        tableCmd.CommandText = "DELETE FROM drinking_water WHERE Id = @id";
-                        tableCmd.Parameters.AddWithValue("@id", recordId);
-                        int rowCount = tableCmd.ExecuteNonQuery();
-
-                        if (rowCount == 0)
-                        {
-                            Console.WriteLine($"ID {recordId} not found!\n");
-                        }
-                        else
-                        {
-                            Console.WriteLine($"Record with ID {recordId} has been deleted.\n");
-                        }
-
-                        connection.Close();
-                    }
-                }
-                catch (SqliteException ex)
-                {
-                    Console.WriteLine("Unexpected Database Error");
-                }
-                catch (Exception exception)
-                {
-                    Console.WriteLine("Unexpected Error");
-                }
-
-                Console.WriteLine("press y to continue, any other key to return to menu");
-                input = Console.ReadLine();
-
-                if (input == null || input == "y") continue;
-                else
-                {
-                    keepDeleting = false;
-                    break;
-                }
-
-
-            }
-        }
-
-        private static void Update()
-        {
-            while (true) // loop until user updates something or exits
-            {
-                PrintAllRecords();
-
-                Console.WriteLine(
-                    "Enter the ID of the record you want to update.\n" +
-                    "Type 'x' to return to the menu."
-                );
-
-                string? input = Console.ReadLine();
-                if (input == null) continue;
-
-                if (input.ToLower() == "x") return; // exit safely
-
-
-                int recordId;
-                if (!int.TryParse(input, out recordId))
-                {
-                    Console.WriteLine("Invalid input. Please enter a numeric ID or 'x' to exit.");
-                    continue;
-                }
-                try
-                {
-                    using (var connection = new SqliteConnection(connectionString))
-                    {
-                        connection.Open();
-
-
-                        var checkCmd = connection.CreateCommand();
-                        checkCmd.CommandText = "SELECT EXISTS(SELECT 1 FROM drinking_water WHERE Id = @id)";
-                        checkCmd.Parameters.AddWithValue("@id", recordId);
-
-                        int exists = Convert.ToInt32(checkCmd.ExecuteScalar());
-                        if (exists == 0)
-                        {
-                            Console.WriteLine($"\nRecord with ID {recordId} does not exist.");
-                            continue;
-                        }
-
-
-                        string? date = GetDateInput();
-                        string habit = GetHabit();
-                        int qty = GetNumberInput("Enter new Quantity:");
-
-
-                        var updateCmd = connection.CreateCommand();
-                        updateCmd.CommandText =
-    "UPDATE drinking_water SET date = @date, HabitType = @habit, Quantity = @qty WHERE Id = @id";
-
-                        updateCmd.Parameters.AddWithValue("@date", date);
-                        updateCmd.Parameters.AddWithValue("@habit", habit);
-                        updateCmd.Parameters.AddWithValue("@qty", qty);
-                        updateCmd.Parameters.AddWithValue("@id", recordId);
-                        updateCmd.ExecuteNonQuery();
-                        connection.Close();
-                    }
-                }
-                catch (SqliteException ex)
-                {
-                    Console.WriteLine("Unexpected Database Error");
-                }
-                catch (Exception exception)
-                {
-                    Console.WriteLine("Unexpected Error");
-                }
-
-                Console.WriteLine("Record updated successfully.");
-                Console.WriteLine("Press any key to return to menu...");
-                Console.ReadKey(true);
-                return;
-            }
-        }
+        Service.GetUserInput();
     }
 }
